Add simple dropdowns
Simple single selection dropdowns. The configuration is messy and the
implementation is not finished, but the core functionality is there.
Keyboard interaction works, auto scroll into view works. I'd like to add
static dropdowns to see if there is a performance difference when not
using Dynamic on larger option sets.