Separate Flags, revamp Image/Icon implementations
'contentImage' is gone in favour of a single function which can be
given either 'Img' (a complete specification for the img element) or
content (which should contain an 'img' widget). The story of which parts
need the right classes is somewhat of a mess in Semantic-UI and I'm
trying to make it easy to just do the right thing, but unfortunately it
takes some experimenation on the part of the user in order for things to
appear correctly in certain situations.

For the modules which embedded 'Icon's and 'Image's (except message),
these config fields are replaced by generic "pre content" widgets. The
user can just use 'icon' and 'image' instead. For messages, only icons
seem to make sense (since it adds the "icon" class).

Removed a bunch of old code in 'Dropdown' (search dropdowns), this will
need replacing eventually. Improved docs in some places.