Fix Transition module's runAction
The new implementation should be more stable on systems with poor
performance. Previously, events could come out of order (due to
performEventAsync) and result in the temporary transitioning states
coming after the final state, leaving the element permanently in a
transitioning state. This essentially resulted in "visibility: hidden"
being stuck on without "display: none" (thus the element could prevent
interaction with things beneath it), which in some cases could
completely break a UI.